requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
Django==5.2.8
django-cloudinary-storage==0.3.0
django-crispy-forms==2.5
django-extensions==4.1
gunicorn==23.0.0
idna==3.11
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.11.0
